Finish Web connection, add creation dialog
