+++ app/src/main/AndroidManifest.xml
@@ -29,11 +29,11 @@
         </activity>
         <activity
                 android:screenOrientation="portrait"
-                android:name=".peer.BlackPeer">
+                android:name=".peer.server.BlackPeer">
         </activity>
         <activity
                 android:screenOrientation="portrait"
-                android:name=".peer.WhitePeer">
+                android:name=".peer.client.WhitePeer">
         </activity>
     </application>
