--- app/src/main/java/com/example/gobang/peer/chessboard/ChessPlaceHandler.java
@@ -1,5 +1,0 @@
-package com.example.gobang.peer.chessboard;
-
-public interface ChessPlaceHandler {
-    void run(byte x, byte y);
-}
