+++ app/src/main/java/com/example/gobang/web/IPAddressUtil.java
@@ -1,33 +1,12 @@
-package com.example.gobang.websocket;
-
-import android.content.Context;
-import android.net.wifi.WifiInfo;
-import android.net.wifi.WifiManager;
-import android.util.Log;
+package com.example.gobang.web;
 
 import java.net.InetAddress;
 import java.net.NetworkInterface;
-import java.net.SocketException;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
 
 public class IPAddressUtil {
-    public static String getWIFIAddress(Context context) {
-        WifiManager wifiManager = (WifiManager) context.getApplicationContext().getSystemService(Context.WIFI_SERVICE);
-        WifiInfo wifiInfo = wifiManager.getConnectionInfo();
-        int ipAddress = wifiInfo.getIpAddress();
-
-        // Convert the IP address from int to string format
-        String ipString = String.format(
-                "%d.%d.%d.%d",
-                (ipAddress & 0xff),
-                (ipAddress >> 8 & 0xff),
-                (ipAddress >> 16 & 0xff),
-                (ipAddress >> 24 & 0xff)
-        );
-        return ipString;
-    }
     public static List<String> getIPAddress() throws Exception{
         List<String> temp=new ArrayList<String>();
         Enumeration<NetworkInterface> networkInterfaces = NetworkInterface.getNetworkInterfaces();
@@ -36,7 +15,7 @@ public class IPAddressUtil {
             Enumeration<InetAddress> inetAddresses = networkInterface.getInetAddresses();
             while (inetAddresses.hasMoreElements()) {
                 InetAddress inetAddress = inetAddresses.nextElement();
-                if (!inetAddress.isLoopbackAddress() && inetAddress.getAddress().length == 4) {
+                if (inetAddress.getAddress().length == 4) {
                     String ipAddress = inetAddress.getHostAddress();
                     temp.add(ipAddress);
                 }
