+++ app/src/main/res/layout/peer.xml
@@ -31,7 +31,7 @@
             app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent" app:layout_constraintHorizontal_bias="0.126"
             app:layout_constraintVertical_bias="0.146" android:textSize="20dp"/>
-    <com.example.gobang.peer.chessboard.ChessBoardASync
+    <com.example.gobang.peer.ChessBoardASync
             android:layout_width="385dp"
             android:layout_height="385dp" app:srcCompat="@drawable/chessboard" android:id="@+id/imageView"
             app:layout_constraintEnd_toEndOf="parent"
